Huge refactor of RHINO (split into multiple modules).
