+++ osl/source_recon/beamforming.py
@@ -11,12 +11,37 @@ import os
 import os.path as op
 
 import numpy as np
-import mne
 from mne import (
     read_forward_solution,
     Covariance,
     compute_covariance,
     compute_raw_covariance,
+)
+from mne.io.meas_info import _simplify_info
+from mne.io.pick import pick_channels_cov, pick_info
+from mne.io.proj import make_projector
+from mne.rank import compute_rank
+from mne.minimum_norm.inverse import _check_depth, _prepare_forward, _get_vertno
+from mne.source_estimate import _get_src_type
+from mne.forward import _subject_from_forward
+from mne.forward.forward import is_fixed_orient
+from mne.beamformer._compute_beamformer import (
+    _reduce_leadfield_rank,
+    Beamformer,
+    _sym_inv_sm,
+)
+from mne.utils import (
+    _check_one_ch_type,
+    _check_info_inv,
+    _check_option,
+    _reg_pinv,
+    _pl,
+    _sym_mat_pow,
+    _check_src_normal,
+    check_version,
+    logger,
+    verbose,
+    warn,
 )
 
 import osl.source_recon.rhino.utils as rhino_utils
@@ -195,7 +220,7 @@ def make_lcmv(
             noise_cov_diag, data_cov.ch_names, bads, data.info["projs"], nfree=1e10
         )
 
-    filters = rhino_utils._make_lcmv(
+    filters = _make_lcmv(
         data.info,
         fwd,
         data_cov,
@@ -357,7 +382,7 @@ def transform_recon_timeseries(
 
         # Sample reference_brain to the desired resolution
         reference_brain_resampled = op.join(
-            coreg_filenames["basefilename"],
+            coreg_filenames["basedir"],
             "MNI152_T1_{}mm_brain.nii.gz".format(spatial_resolution),
         )
 
@@ -415,3 +440,480 @@ def transform_recon_timeseries(
             recon_indices[cc] = recon_index
 
     return recon_timeseries_out, reference_brain_resampled, coords_out, recon_indices
+
+
+@verbose
+def _make_lcmv(
+    info,
+    forward,
+    data_cov,
+    reg=0.05,
+    noise_cov=None,
+    label=None,
+    pick_ori=None,
+    rank="info",
+    noise_rank="info",
+    weight_norm="unit-noise-gain-invariant",
+    reduce_rank=False,
+    depth=None,
+    inversion="matrix",
+    verbose=None,
+):
+    """Compute LCMV spatial filter.
+
+    RHINO version of mne.beamformer.make_lcmv Code that is different to
+    mne.beamformer.make_lcmv is labelled with MWW.
+    """
+    # check number of sensor types present in the data and ensure a noise cov
+    info = _simplify_info(info)
+    noise_cov, _, allow_mismatch = _check_one_ch_type(
+        "lcmv", info, forward, data_cov, noise_cov
+    )
+    # XXX we need this extra picking step (can't just rely on minimum norm's
+    # because there can be a mismatch. Should probably add an extra arg to
+    # _prepare_beamformer_input at some point (later)
+    picks = _check_info_inv(info, forward, data_cov, noise_cov)
+    info = pick_info(info, picks)
+
+    data_rank = compute_rank(data_cov, rank=rank, info=info)
+    noise_rank = compute_rank(noise_cov, rank=noise_rank, info=info)
+
+    if False:  # MWW
+        for key in data_rank:
+            if (
+                key not in noise_rank or data_rank[key] != noise_rank[key]
+            ) and not allow_mismatch:
+                raise ValueError(
+                    "%s data rank (%s) did not match the noise "
+                    "rank (%s)" % (key, data_rank[key], noise_rank.get(key, None))
+                )
+    # MWW
+    # del noise_rank
+    rank = data_rank
+    logger.info("Making LCMV beamformer with data cov rank %s" % (rank,))
+    # MWW added:
+    logger.info("Making LCMV beamformer with noise cov rank %s" % (noise_rank,))
+
+    del data_rank
+    depth = _check_depth(depth, "depth_sparse")
+    if inversion == "single":
+        depth["combine_xyz"] = False
+
+    # MWW
+    (
+        is_free_ori, info, proj, vertno, G, whitener, nn, orient_std
+    ) = _prepare_beamformer_input(
+        info, forward, label, pick_ori,
+        noise_cov=noise_cov, rank=noise_rank, pca=False, **depth,
+    )
+
+    ch_names = list(info["ch_names"])
+
+    data_cov = pick_channels_cov(data_cov, include=ch_names)
+    Cm = data_cov._get_square()
+    if "estimator" in data_cov:
+        del data_cov["estimator"]
+    rank_int = sum(rank.values())
+    del rank
+
+    # compute spatial filter
+    n_orient = 3 if is_free_ori else 1
+    W, max_power_ori = _compute_beamformer(
+        G, Cm, reg, n_orient, weight_norm, pick_ori, reduce_rank, rank_int,
+        inversion=inversion, nn=nn, orient_std=orient_std, whitener=whitener,
+    )
+
+    # get src type to store with filters for _make_stc
+    src_type = _get_src_type(forward["src"], vertno)
+
+    # get subject to store with filters
+    subject_from = _subject_from_forward(forward)
+
+    # Is the computed beamformer a scalar or vector beamformer?
+    is_free_ori = is_free_ori if pick_ori in [None, "vector"] else False
+    is_ssp = bool(info["projs"])
+
+    filters = Beamformer(
+        kind="LCMV",
+        weights=W,
+        data_cov=data_cov,
+        noise_cov=noise_cov,
+        whitener=whitener,
+        weight_norm=weight_norm,
+        pick_ori=pick_ori,
+        ch_names=ch_names,
+        proj=proj,
+        is_ssp=is_ssp,
+        vertices=vertno,
+        is_free_ori=is_free_ori,
+        n_sources=forward["nsource"],
+        src_type=src_type,
+        source_nn=forward["source_nn"].copy(),
+        subject=subject_from,
+        rank=rank_int,
+        max_power_ori=max_power_ori,
+        inversion=inversion,
+    )
+
+    return filters
+
+
+def _compute_beamformer(
+    G, Cm, reg, n_orient, weight_norm, pick_ori, reduce_rank, rank, inversion, nn,
+    orient_std, whitener
+):
+    """Compute a spatial beamformer filter (LCMV or DICS).
+
+    For more detailed information on the parameters, see the docstrings of
+    `make_lcmv` and `make_dics`.
+
+    RHINO version of mne.beamformer._compute_beamformer
+    See lines marked MWW for where code has been changed
+
+    Parameters
+    ----------
+    G : ndarray, shape (n_dipoles, n_channels)
+        The leadfield.
+    Cm : ndarray, shape (n_channels, n_channels)
+        The data covariance matrix.
+    reg : float
+        Regularization parameter.
+    n_orient : int
+        Number of dipole orientations defined at each source point
+    weight_norm : None | 'unit-noise-gain' | 'nai'
+        The weight normalization scheme to use.
+    pick_ori : None | 'normal' | 'max-power' | max-power-pre-weight-norm
+        The source orientation to compute the beamformer in.
+    reduce_rank : bool
+        Whether to reduce the rank by one during computation of the filter.
+    rank : dict | None | 'full' | 'info'
+        See compute_rank.
+    inversion : 'matrix' | 'single'
+        The inversion scheme to compute the weights.
+    nn : ndarray, shape (n_dipoles, 3)
+        The source normals.
+    orient_std : ndarray, shape (n_dipoles,)
+        The std of the orientation prior used in weighting the lead fields.
+    whitener : ndarray, shape (n_channels, n_channels)
+        The whitener.
+
+    Returns
+    -------
+    W : ndarray, shape (n_dipoles, n_channels)
+        The beamformer filter weights.
+    """
+    _check_option(
+        "weight_norm",
+        weight_norm,
+        ["unit-noise-gain-invariant", "unit-noise-gain", "nai", None],
+    )
+
+    # Whiten the data covariance
+    Cm = whitener @ Cm @ whitener.T.conj()
+    # Restore to properly Hermitian as large whitening coefs can have bad
+    # rounding error
+
+    Cm[:] = (Cm + Cm.T.conj()) / 2.0
+
+    assert Cm.shape == (G.shape[0],) * 2
+    s, _ = np.linalg.eigh(Cm)
+    if not (s >= -s.max() * 1e-7).all():
+        # This shouldn't ever happen, but just in case
+        warn(
+            "data covariance does not appear to be positive semidefinite, "
+            "results will likely be incorrect"
+        )
+    # Tikhonov regularization using reg parameter to control for
+    # trade-off between spatial resolution and noise sensitivity
+    # eq. 25 in Gross and Ioannides, 1999 Phys. Med. Biol. 44 2081
+    Cm_inv, loading_factor, rank = _reg_pinv(Cm, reg, rank)
+
+    assert orient_std.shape == (G.shape[1],)
+    n_sources = G.shape[1] // n_orient
+    assert nn.shape == (n_sources, 3)
+
+    logger.info(
+        "Computing beamformer filters for %d source%s" % (n_sources, _pl(n_sources))
+    )
+    n_channels = G.shape[0]
+    assert n_orient in (3, 1)
+    Gk = np.reshape(G.T, (n_sources, n_orient, n_channels)).transpose(0, 2, 1)
+    assert Gk.shape == (n_sources, n_channels, n_orient)
+    sk = np.reshape(orient_std, (n_sources, n_orient))
+    del G, orient_std
+    pinv_kwargs = dict()
+    if check_version("numpy", "1.17"):
+        pinv_kwargs["hermitian"] = True
+
+    _check_option("reduce_rank", reduce_rank, (True, False))
+
+    # inversion of the denominator
+    _check_option("inversion", inversion, ("matrix", "single"))
+    if (
+        inversion == "single"
+        and n_orient > 1
+        and pick_ori == "vector"
+        and weight_norm == "unit-noise-gain-invariant"
+    ):
+        raise ValueError(
+            'Cannot use pick_ori="vector" with inversion="single" and '
+            'weight_norm="unit-noise-gain-invariant"'
+        )
+    if reduce_rank and inversion == "single":
+        raise ValueError(
+            'reduce_rank cannot be used with inversion="single"; '
+            'consider using inversion="matrix" if you have a '
+            "rank-deficient forward model (i.e., from a sphere "
+            "model with MEG channels), otherwise consider using "
+            "reduce_rank=False"
+        )
+    if n_orient > 1:
+        _, Gk_s, _ = np.linalg.svd(Gk, full_matrices=False)
+        assert Gk_s.shape == (n_sources, n_orient)
+        if not reduce_rank and (Gk_s[:, 0] > 1e6 * Gk_s[:, 2]).any():
+            raise ValueError(
+                "Singular matrix detected when estimating spatial filters. "
+                "Consider reducing the rank of the forward operator by using "
+                "reduce_rank=True."
+            )
+        del Gk_s
+
+    # ------------------------------------------------------------------
+    # 1. Reduce rank of the lead field
+    if reduce_rank:
+        Gk = _reduce_leadfield_rank(Gk)
+
+    def _compute_bf_terms(Gk, Cm_inv):
+        bf_numer = np.matmul(Gk.swapaxes(-2, -1).conj(), Cm_inv)
+        bf_denom = np.matmul(bf_numer, Gk)
+        return bf_numer, bf_denom
+
+    # ------------------------------------------------------------------
+    # 2. Reorient lead field in direction of max power or normal
+    if pick_ori == "max-power" or pick_ori == "max-power-pre-weight-norm":
+        assert n_orient == 3
+        _, bf_denom = _compute_bf_terms(Gk, Cm_inv)
+
+        if pick_ori == "max-power":
+            if weight_norm is None:
+                ori_numer = np.eye(n_orient)[np.newaxis]
+                ori_denom = bf_denom
+            else:
+                # compute power, cf Sekihara & Nagarajan 2008, eq. 4.47
+                ori_numer = bf_denom
+                # Cm_inv should be Hermitian so no need for .T.conj()
+                ori_denom = np.matmul(
+                    np.matmul(Gk.swapaxes(-2, -1).conj(), Cm_inv @ Cm_inv), Gk
+                )
+
+            ori_denom_inv = _sym_inv_sm(ori_denom, reduce_rank, inversion, sk)
+            ori_pick = np.matmul(ori_denom_inv, ori_numer)
+
+        # MWW
+        else:  # pick_ori == 'max-power-pre-weight-norm':
+
+            # Compute power, see eq 5 in Brookes et al, Optimising experimental
+            # design for MEG beamformer imaging, Neuroimage 2008
+            # This optimises the orientation by maximising the power
+            # BEFORE any weight normalisation is performed
+            ori_pick = _sym_inv_sm(bf_denom, reduce_rank, inversion, sk)
+
+        assert ori_pick.shape == (n_sources, n_orient, n_orient)
+
+        # pick eigenvector that corresponds to maximum eigenvalue:
+        eig_vals, eig_vecs = np.linalg.eig(ori_pick.real)  # not Hermitian!
+        # sort eigenvectors by eigenvalues for picking:
+        order = np.argsort(np.abs(eig_vals), axis=-1)
+        # eig_vals = np.take_along_axis(eig_vals, order, axis=-1)
+        max_power_ori = eig_vecs[np.arange(len(eig_vecs)), :, order[:, -1]]
+        assert max_power_ori.shape == (n_sources, n_orient)
+
+        # set the (otherwise arbitrary) sign to match the normal
+        signs = np.sign(np.sum(max_power_ori * nn, axis=1, keepdims=True))
+        signs[signs == 0] = 1.0
+        max_power_ori *= signs
+
+        # Compute the lead field for the optimal orientation,
+        # and adjust numer/denom
+
+        Gk = np.matmul(Gk, max_power_ori[..., np.newaxis])
+
+        n_orient = 1
+    else:
+        max_power_ori = None
+        if pick_ori == "normal":
+            Gk = Gk[..., 2:3]
+            n_orient = 1
+
+    # ----------------------------------------------------------------------
+    # 3. Compute numerator and denominator of beamformer formula (unit-gain)
+
+    bf_numer, bf_denom = _compute_bf_terms(Gk, Cm_inv)
+    assert bf_denom.shape == (n_sources,) + (n_orient,) * 2
+    assert bf_numer.shape == (n_sources, n_orient, n_channels)
+    del Gk  # lead field has been adjusted and should not be used anymore
+
+    # ----------------------------------------------------------------------
+    # 4. Invert the denominator
+
+    # Here W is W_ug, i.e.:
+    # G.T @ Cm_inv / (G.T @ Cm_inv @ G)
+    bf_denom_inv = _sym_inv_sm(bf_denom, reduce_rank, inversion, sk)
+    assert bf_denom_inv.shape == (n_sources, n_orient, n_orient)
+    W = np.matmul(bf_denom_inv, bf_numer)
+    assert W.shape == (n_sources, n_orient, n_channels)
+    del bf_denom_inv, sk
+
+    # ----------------------------------------------------------------------
+    # 5. Re-scale filter weights according to the selected weight_norm
+
+    # Weight normalization is done by computing, for each source::
+    #
+    #     W_ung = W_ug / sqrt(W_ug @ W_ug.T)
+    #
+    # with W_ung referring to the unit-noise-gain (weight normalized) filter
+    # and W_ug referring to the above-calculated unit-gain filter stored in W.
+
+    if weight_norm is not None:
+        # Three different ways to calculate the normalization factors here.
+        # Only matters when in vector mode, as otherwise n_orient == 1 and
+        # they are all equivalent. Sekihara 2008 says to use
+        #
+        # In MNE < 0.21, we just used the Frobenius matrix norm:
+        #
+        #    noise_norm = np.linalg.norm(W, axis=(1, 2), keepdims=True)
+        #    assert noise_norm.shape == (n_sources, 1, 1)
+        #    W /= noise_norm
+        #
+        # Sekihara 2008 says to use sqrt(diag(W_ug @ W_ug.T)), which is not
+        # rotation invariant:
+        if weight_norm in ("unit-noise-gain", "nai"):
+            noise_norm = np.matmul(W, W.swapaxes(-2, -1).conj()).real
+            noise_norm = np.reshape(  # np.diag operation over last two axes
+                noise_norm, (n_sources, -1, 1)
+            )[:, :: n_orient + 1]
+            np.sqrt(noise_norm, out=noise_norm)
+            noise_norm[noise_norm == 0] = np.inf
+            assert noise_norm.shape == (n_sources, n_orient, 1)
+            W /= noise_norm
+        else:
+            assert weight_norm == "unit-noise-gain-invariant"
+            # Here we use sqrtm. The shortcut:
+            #
+            #    use = W
+            #
+            # ... does not match the direct route (it is rotated!), so we'll
+            # use the direct one to match FieldTrip:
+            use = bf_numer
+            inner = np.matmul(use, use.swapaxes(-2, -1).conj())
+            W = np.matmul(_sym_mat_pow(inner, -0.5), use)
+            noise_norm = 1.0
+
+        if weight_norm == "nai":
+            # Estimate noise level based on covariance matrix, taking the
+            # first eigenvalue that falls outside the signal subspace or the
+            # loading factor used during regularization, whichever is largest.
+            if rank > len(Cm):
+                # Covariance matrix is full rank, no noise subspace!
+                # Use the loading factor as noise ceiling.
+                if loading_factor == 0:
+                    raise RuntimeError(
+                        "Cannot compute noise subspace with a full-rank "
+                        "covariance matrix and no regularization. Try "
+                        "manually specifying the rank of the covariance "
+                        "matrix or using regularization."
+                    )
+                noise = loading_factor
+            else:
+                noise, _ = np.linalg.eigh(Cm)
+                noise = noise[-rank]
+                noise = max(noise, loading_factor)
+            W /= np.sqrt(noise)
+
+    W = W.reshape(n_sources * n_orient, n_channels)
+    logger.info("Filter computation complete")
+    return W, max_power_ori
+
+
+def _prepare_beamformer_input(
+    info,
+    forward,
+    label=None,
+    pick_ori=None,
+    noise_cov=None,
+    rank=None,
+    pca=False,
+    loose=None,
+    combine_xyz="fro",
+    exp=None,
+    limit=None,
+    allow_fixed_depth=True,
+    limit_depth_chs=False,
+):
+    """Input preparation common for LCMV, DICS, and RAP-MUSIC.
+
+    RHINO version of mne.beamformer._prepare_beamformer_input.
+
+    See lines marked MWW (or CG) for where code has been changed.
+    """
+
+    # MWW
+    # _check_option('pick_ori', pick_ori, ('normal', 'max-power', 'vector', None))
+    _check_option(
+        "pick_ori", pick_ori,
+        ("normal", "max-power", "vector", "max-power-pre-weight-norm", None),
+    )
+
+    # CG
+    # Restrict forward solution to selected vertices
+    #if label is not None:
+    #    _, src_sel = label_src_vertno_sel(label, forward["src"])
+    #    forward = _restrict_forward_to_src_sel(forward, src_sel)
+
+    if loose is None:
+        loose = 0.0 if is_fixed_orient(forward) else 1.0
+
+    # CG
+    #if noise_cov is None:
+    #    noise_cov = make_ad_hoc_cov(info, std=1.0)
+
+    forward, info_picked, gain, _, orient_prior, _, trace_GRGT, noise_cov, whitener = \
+        _prepare_forward(
+            forward, info, noise_cov, "auto", loose, rank=rank, pca=pca, use_cps=True,
+            exp=exp, limit_depth_chs=limit_depth_chs, combine_xyz=combine_xyz,
+            limit=limit, allow_fixed_depth=allow_fixed_depth,
+        )
+    is_free_ori = not is_fixed_orient(forward)  # could have been changed
+    nn = forward["source_nn"]
+    if is_free_ori:  # take Z coordinate
+        nn = nn[2::3]
+    nn = nn.copy()
+    vertno = _get_vertno(forward["src"])
+    if forward["surf_ori"]:
+        nn[...] = [0, 0, 1]  # align to local +Z coordinate
+    if pick_ori is not None and not is_free_ori:
+        raise ValueError(
+            "Normal or max-power orientation (got %r) can only be picked when "
+            "a forward operator with free orientation is used." % (pick_ori,)
+        )
+    if pick_ori == "normal" and not forward["surf_ori"]:
+        raise ValueError(
+            "Normal orientation can only be picked when a "
+            "forward operator oriented in surface coordinates is "
+            "used."
+        )
+    _check_src_normal(pick_ori, forward["src"])
+    del forward, info
+
+    # Undo the scaling that MNE prefers
+    scale = np.sqrt((noise_cov["eig"] > 0).sum() / trace_GRGT)
+    gain /= scale
+    if orient_prior is not None:
+        orient_std = np.sqrt(orient_prior)
+    else:
+        orient_std = np.ones(gain.shape[1])
+
+    # Get the projector
+    proj, _, _ = make_projector(info_picked["projs"], info_picked["ch_names"])
+
+    return is_free_ori, info_picked, proj, vertno, gain, whitener, nn, orient_std
