+++ osl/source_recon/rhino/__init__.py
@@ -5,5 +5,9 @@
 """
 
 from .rhino import *  # noqa: F401, F403
+from .polhemus import *  # noqa: F401, F403
+from .freesurfer import *  # noqa: F401, F403
+from .forward_model import *  # noqa: F401, F403
+from .surfaces import *  # noqa: F401, F403
 from .fsl_wrappers import *  # noqa: F401, F403
 from . import utils  # noqa: F401, F403
