+++ osl/source_recon/rhino/utils.py
@@ -8,54 +8,26 @@
 #          Chetan Gohil <chetan.gohil@psych.ox.ac.uk>
 
 import os
-import os.path as op
 
 import numpy as np
 from sklearn.neighbors import NearestNeighbors
 import nibabel as nib
 import pandas as pd
-import matplotlib.pyplot as plt
 from scipy import LowLevelCallable
 from scipy.ndimage import generic_filter
 from scipy.spatial import KDTree
 
-from mne import write_surface, Transform
+from mne import Transform
 from mne.transforms import read_trans
-from mne.io.meas_info import _simplify_info
-from mne.io.pick import pick_channels_cov, pick_info
-from mne.io.proj import make_projector, Projection
-from mne.source_estimate import _get_src_type
-from mne.forward import _subject_from_forward
-from mne.forward.forward import is_fixed_orient
-from mne.utils import (
-    verbose,
-    _check_one_ch_type,
-    _check_info_inv,
-    _reg_pinv,
-    _check_option,
-    logger,
-    _check_src_normal,
-    _pl,
-    check_version,
-    _sym_mat_pow,
-    warn,
-)
-from mne.rank import compute_rank
-from mne.minimum_norm.inverse import _check_depth, _prepare_forward, _get_vertno
-from mne.beamformer._compute_beamformer import (
-    _reduce_leadfield_rank,
-    Beamformer,
-    _sym_inv_sm,
-)
 
 from numba import cfunc, carray
 from numba.types import intc, intp, float64, voidptr
 from numba.types import CPointer
 
-from osl.utils import soft_import
-
 import logging
 logging.getLogger("numba").setLevel(logging.WARNING)
+
+from osl.source_recon.beamforming import transform_recon_timeseries
 
 
 def system_call(cmd, verbose=False):
@@ -476,8 +448,6 @@ def icp(A, B, init_pose=None, max_iterations=50, tolerance=0.0001):
     From: https://github.com/ClayFlannigan/icp/blob/master/icp.py
     """
 
-    # assert A.shape == B.shape
-
     # get number of dimensions
     m = A.shape[1]
 
@@ -493,8 +463,6 @@ def icp(A, B, init_pose=None, max_iterations=50, tolerance=0.0001):
 
     prev_error = 0
 
-    # import pdb; pdb.pdb.set_trace()
-
     kdtree = KDTree(dst[:m, :].T)
 
     for i in range(max_iterations):
@@ -503,9 +471,8 @@ def icp(A, B, init_pose=None, max_iterations=50, tolerance=0.0001):
 
         distances, indices = kdtree.query(src[:m, :].T)
 
-        # import pdb; pdb.pdb.set_trace()
-
-        # compute the transformation between the current source and nearest destination points
+        # compute the transformation between the current source and nearest
+        # destination points
         T = best_fit_transform(src[:m, :].T, dst[:m, indices].T)
 
         # update the current source
@@ -644,95 +611,6 @@ def rhino_icp(smri_headshape_polhemus, polhemus_headshape_polhemus, Ninits=10):
     return xform, err, err_old
 
 
-def create_freesurfer_mesh(
-    infile, surf_outfile, xform_mri_voxel2mri, nii_mesh_file=None
-):
-    """Creates surface mesh in .surf format and in native mri space in mm
-    from infile.
-
-    Parameters
-    ----------
-    infile : string
-    Either:
-        1) .nii.gz file containing zero's for background and one's for surface
-        2) .vtk file generated by bet_surf (in which case the path to the
-        strutural MRI, smri_file, must be included as an input)
-    surf_outfile : string
-        Path to the .surf file generated, containing the surface
-        mesh in mm
-    xform_mri_voxel2mri : numpy.ndarray
-        4x4 array
-        Transform from voxel indices to native/mri mm
-    nii_mesh_file : string
-        Path to the niftii mesh file that is the niftii equivalent
-        of vtk file passed in as infile (only needed if infile
-        is a vtk file)
-    """
-
-    pth, name = op.split(infile)
-    name, ext = op.splitext(name)
-
-    if ext == ".gz":
-        print("Creating surface mesh for {} .....".format(infile))
-
-        # Soft import raising an informative warning if not installed
-        o3s = soft_import("open3d")
-
-        name, ext = op.splitext(name)
-        if ext != ".nii":
-            raise ValueError("Invalid infile. Needs to be a .nii.gz or .vtk file")
-
-        # convert to point cloud in voxel indices
-        nii_nativeindex = niimask2indexpointcloud(infile)
-
-        # print('Num of vertices to create mesh from = {}'.format(nii_nativeindex.shape[1]))
-
-        step = 1
-        nii_native = xform_points(xform_mri_voxel2mri, nii_nativeindex[:, 0:-1:step])
-
-        pcd = o3d.geometry.PointCloud()
-        pcd.points = o3d.utility.Vector3dVector(nii_native.T)
-        pcd.estimate_normals()
-        # to obtain a consistent normal orientation
-        pcd.orient_normals_towards_camera_location(pcd.get_center())
-
-        # or you might want to flip the normals to make them point outward, not mandatory
-        pcd.normals = o3d.utility.Vector3dVector(-np.asarray(pcd.normals))
-
-        mesh = o3d.geometry.TriangleMesh.create_from_point_cloud_poisson(pcd, depth=8)[0]
-
-        # mesh = mesh.simplify_quadric_decimation(nii_nativeindex.shape[1])
-
-        verts = np.asarray(mesh.vertices)
-        tris = np.asarray(mesh.triangles).astype(int)
-
-        # output in freesurfer file format
-        write_surface(
-            surf_outfile, verts, tris, file_format="freesurfer", overwrite=True
-        )
-    elif ext == ".vtk":
-
-        if nii_mesh_file is None:
-            raise ValueError(
-                "You must specify a nii_mesh_file (niftii format), if \
-infile format is vtk"
-            )
-
-        rrs_native, tris_native = _get_vtk_mesh_native(infile, nii_mesh_file)
-
-        write_surface(
-            surf_outfile,
-            rrs_native,
-            tris_native,
-            file_format="freesurfer",
-            overwrite=True,
-        )
-    else:
-        raise ValueError("Invalid infile. Needs to be a .nii.gz or .vtk file")
-
-    #print('Written new surface mesh: {}'.format(surf_outfile))
-
-
 def _get_vtk_mesh_native(vtk_mesh_file, nii_mesh_file):
     """
     Returns mesh rrs in native space in mm and the meash tris for the passed
@@ -747,7 +625,7 @@ def _get_vtk_mesh_native(vtk_mesh_file, nii_mesh_file):
     num_rrs = int(data.iloc[3, 1])
 
     # these will be in voxel index space
-    rrs_flirtcoords = data.iloc[4 : num_rrs + 4, 0:3].to_numpy().astype(np.float64)
+    rrs_flirtcoords = data.iloc[4: num_rrs + 4, 0:3].to_numpy().astype(np.float64)
 
     # move to from flirtcoords mm to mri mm (native) space
     xform_flirtcoords2nii = _get_flirtcoords2native_xform(nii_mesh_file)
@@ -755,7 +633,7 @@ def _get_vtk_mesh_native(vtk_mesh_file, nii_mesh_file):
 
     num_tris = int(data.iloc[num_rrs + 4, 1])
     tris_nii = (
-        data.iloc[num_rrs + 5 : num_rrs + 5 + num_tris, 1:4].to_numpy().astype(int)
+        data.iloc[num_rrs + 5: num_rrs + 5 + num_tris, 1:4].to_numpy().astype(int)
     )
 
     return rrs_nii, tris_nii
@@ -835,7 +713,7 @@ def _transform_vtk_mesh(
     data = pd.read_csv(vtk_mesh_file_in, delim_whitespace=True)
 
     num_rrs = int(data.iloc[3, 1])
-    data.iloc[4 : num_rrs + 4, 0:3] = rrs_out
+    data.iloc[4: num_rrs + 4, 0:3] = rrs_out
 
     # write new vtk file
     data.to_csv(out_vtk_file, sep=" ", index=False)
@@ -902,613 +780,157 @@ def _get_flirt_xform_between_axes(from_nii, target_nii):
     return from2to
 
 
-def plot_polhemus_points(
-    txt_fnames, colors=None, scales=None, markers=None, alphas=None
+def _timeseries2nii(
+    timeseries, timeseries_coords, reference_mask_fname, out_nii_fname, times=None
 ):
-    plt.figure()
-    ax = plt.axes(projection="3d")
+    """Maps the (ndipoles,tpts) array of timeseries to the grid defined by
+    reference_mask_fname and outputs them as a niftii file.
 
-    for ss in range(len(txt_fnames)):
-
-        if alphas is None:
-            alpha = 1
-        else:
-            alpha = alphas[ss]
-
-        if colors is None:
-            color = (0.5, 0.5, 0.5)
-        else:
-            color = colors[ss]
-
-        if scales is None:
-            scale = 10
-        else:
-            scale = scales[ss]
-
-        if markers is None:
-            marker = 1
-        else:
-            marker = markers[ss]
-
-        pnts = np.loadtxt(txt_fnames[ss])
-
-        ax.scatter(
-            pnts[0], pnts[1], pnts[2], color=color, s=scale, alpha=alpha, marker=marker,
-        )
-
-
-@verbose
-def _make_lcmv(
-    info,
-    forward,
-    data_cov,
-    reg=0.05,
-    noise_cov=None,
-    label=None,
-    pick_ori=None,
-    rank="info",
-    noise_rank="info",
-    weight_norm="unit-noise-gain-invariant",
-    reduce_rank=False,
-    depth=None,
-    inversion="matrix",
-    verbose=None,
-):
-    """Compute LCMV spatial filter.
+    Assumes the timeseries' dipoles correspond to those in reference_mask_fname.
+    Both timeseries and reference_mask_fname are often output from
+    rhino.transform_recon_timeseries.
 
     Parameters
     ----------
-    info : instance of Info
-        The measurement info to specify the channels to include.
-        Bad channels in info['bads'] are not used.
-    forward : instance of Forward
-        Forward operator.
-    data_cov : instance of Covariance
-        The data covariance.
-    reg : float
-        The regularization for the whitened data covariance.
-    noise_cov : instance of Covariance
-        The noise covariance. If provided, whitening will be done. Providing a
-        noise covariance is mandatory if you mix sensor types, e.g.
-        gradiometers with magnetometers or EEG with MEG.
-    label : instance of Label
-        Restricts the LCMV solution to a given label.
+    timeseries : (ndipoles, ntpts) numpy.ndarray
+        Time courses.
+        Assumes the timeseries' dipoles correspond to those in reference_mask_fname.
+        Typically derives from rhino.transform_recon_timeseries
+    timeseries_coords : (ndipoles, 3) numpy.ndarray
+        Coords in mm for dipoles corresponding to passed in timeseries
+    reference_mask_fname : string
+        A nii.gz mask file name
+        (with zero for background, and !=0 for the mask)
+        Assumes the mask was used to set dipoles for timeseries,
+        typically derived from rhino.transform_recon_timeseries
+    out_nii_fname : string
+        output name of niftii file
+    times : (ntpts, ) numpy.ndarray
+        Times points in seconds.
+        Assume that times are regularly spaced.
+        Used to set nii file up correctly.
 
     Returns
     -------
-    filters : instance of Beamformer
-        Dictionary containing filter weights from LCMV beamformer.
-        Contains the following keys:
-            'kind' : str
-                The type of beamformer, in this case 'LCMV'.
-            'weights' : array
-                The filter weights of the beamformer.
-            'data_cov' : instance of Covariance
-                The data covariance matrix used to compute the beamformer.
-            'noise_cov' : instance of Covariance | None
-                The noise covariance matrix used to compute the beamformer.
-            'whitener' : None | ndarray, shape (n_channels, n_channels)
-                Whitening matrix, provided if whitening was applied to the
-                covariance matrix and leadfield during computation of the
-                beamformer weights.
-            'weight_norm' : str | None
-                Type of weight normalization used to compute the filter
-                weights.
-            'pick-ori' : None | 'max-power' | 'normal' | 'vector'
-                The orientation in which the beamformer filters were computed.
-            'ch_names' : list of str
-                Channels used to compute the beamformer.
-            'proj' : array
-                Projections used to compute the beamformer.
-            'is_ssp' : bool
-                If True, projections were applied prior to filter computation.
-            'vertices' : list
-                Vertices for which the filter weights were computed.
-            'is_free_ori' : bool
-                If True, the filter was computed with free source orientation.
-            'n_sources' : int
-                Number of source location for which the filter weight were
-                computed.
-            'src_type' : str
-                Type of source space.
-            'source_nn' : ndarray, shape (n_sources, 3)
-                For each source location, the surface normal.
-            'proj' : ndarray, shape (n_channels, n_channels)
-                Projections used to compute the beamformer.
-            'subject' : str
-                The subject ID.
-            'rank' : int
-                The rank of the data covariance matrix used to compute the
-                beamformer weights.
-            'max-power-ori' : ndarray, shape (n_sources, 3) | None
-                When pick_ori='max-power', this fields contains the estimated
-                direction of maximum power at each source location.
-            'inversion' : 'single' | 'matrix'
-                Whether the spatial filters were computed for each dipole
-                separately or jointly for all dipoles at each vertex using a
-                matrix inversion.
-
-    Notes
-    -----
-    RHINO version of mne.beamformer.make_lcmv
-
-    Code that is different to mne.beamformer.make_lcmv is labelled with MWW.
-
-    The original reference is :footcite:`VanVeenEtAl1997`.
-
-    To obtain the Sekihara unit-noise-gain vector beamformer, you should use
-    ``weight_norm='unit-noise-gain', pick_ori='vector'`` followed by
-    :meth:`vec_stc.project('pca', src) <mne.VectorSourceEstimate.project>`.
-
-    .. versionchanged:: 0.21
-       The computations were extensively reworked, and the default for
-       ``weight_norm`` was set to ``'unit-noise-gain-invariant'``.
-
-    References
-    ----------
-    .. footbibliography::
+    out_nii_fname : string
+        Name of output niftii file
     """
-    # check number of sensor types present in the data and ensure a noise cov
-    info = _simplify_info(info)
-    noise_cov, _, allow_mismatch = _check_one_ch_type(
-        "lcmv", info, forward, data_cov, noise_cov
-    )
-    # XXX we need this extra picking step (can't just rely on minimum norm's
-    # because there can be a mismatch. Should probably add an extra arg to
-    # _prepare_beamformer_input at some point (later)
-    picks = _check_info_inv(info, forward, data_cov, noise_cov)
-    info = pick_info(info, picks)
 
-    data_rank = compute_rank(data_cov, rank=rank, info=info)
-    noise_rank = compute_rank(noise_cov, rank=noise_rank, info=info)
+    if len(timeseries.shape) == 1:
+        timeseries = np.reshape(timeseries, [-1, 1])
 
-    if False:  # MWW
-        for key in data_rank:
-            if (
-                key not in noise_rank or data_rank[key] != noise_rank[key]
-            ) and not allow_mismatch:
-                raise ValueError(
-                    "%s data rank (%s) did not match the noise "
-                    "rank (%s)" % (key, data_rank[key], noise_rank.get(key, None))
-                )
-    # MWW
-    # del noise_rank
-    rank = data_rank
-    logger.info("Making LCMV beamformer with data cov rank %s" % (rank,))
-    # MWW added:
-    logger.info("Making LCMV beamformer with noise cov rank %s" % (noise_rank,))
+    mni_nii_nib = nib.load(reference_mask_fname)
+    coords_ind = niimask2indexpointcloud(reference_mask_fname).T
+    coords_mni, tmp = niimask2mmpointcloud(reference_mask_fname)
 
-    del data_rank
-    depth = _check_depth(depth, "depth_sparse")
-    if inversion == "single":
-        depth["combine_xyz"] = False
+    mni_nii_values = mni_nii_nib.get_fdata()
+    mni_nii_values = np.zeros(np.append(mni_nii_values.shape, timeseries.shape[1]))
 
-    # MWW
-    is_free_ori, info, proj, vertno, G, whitener, nn, orient_std = \
-         _prepare_beamformer_input(
-            info, forward, label, pick_ori, noise_cov=noise_cov, rank=noise_rank,
-            pca=False, **depth,
-        )
+    kdtree = KDTree(coords_mni.T)
+    gridstep = int(get_gridstep(coords_mni.T) / 1000)
 
-    ch_names = list(info["ch_names"])
+    for ind in range(timeseries_coords.shape[1]):
+        distance, index = kdtree.query(timeseries_coords[:, ind])
+        # Exclude any timeseries_coords that are further than gridstep away
+        # from the best matching coords_mni
+        if distance < gridstep:
+            mni_nii_values[
+                coords_ind[ind, 0], coords_ind[ind, 1], coords_ind[ind, 2], :
+            ] = timeseries[ind, :]
 
-    data_cov = pick_channels_cov(data_cov, include=ch_names)
-    Cm = data_cov._get_square()
-    if "estimator" in data_cov:
-        del data_cov["estimator"]
-    rank_int = sum(rank.values())
-    del rank
+    # SAVE AS NIFTI
+    vol_nii = nib.Nifti1Image(mni_nii_values, mni_nii_nib.affine)
 
-    # compute spatial filter
-    n_orient = 3 if is_free_ori else 1
-    W, max_power_ori = _compute_beamformer(
-        G, Cm, reg, n_orient, weight_norm, pick_ori, reduce_rank, rank_int,
-        inversion=inversion, nn=nn, orient_std=orient_std, whitener=whitener,
-    )
+    vol_nii.header.set_xyzt_units(2)  # mm
+    if times is not None:
+        vol_nii.header["pixdim"][4] = times[1] - times[0]
+        vol_nii.header["toffset"] = -0.5
+        vol_nii.header.set_xyzt_units(2, 8)  # mm and secs
 
-    # get src type to store with filters for _make_stc
-    src_type = _get_src_type(forward["src"], vertno)
+    nib.save(vol_nii, out_nii_fname)
 
-    # get subject to store with filters
-    subject_from = _subject_from_forward(forward)
-
-    # Is the computed beamformer a scalar or vector beamformer?
-    is_free_ori = is_free_ori if pick_ori in [None, "vector"] else False
-    is_ssp = bool(info["projs"])
-
-    filters = Beamformer(
-        kind="LCMV",
-        weights=W,
-        data_cov=data_cov,
-        noise_cov=noise_cov,
-        whitener=whitener,
-        weight_norm=weight_norm,
-        pick_ori=pick_ori,
-        ch_names=ch_names,
-        proj=proj,
-        is_ssp=is_ssp,
-        vertices=vertno,
-        is_free_ori=is_free_ori,
-        n_sources=forward["nsource"],
-        src_type=src_type,
-        source_nn=forward["source_nn"].copy(),
-        subject=subject_from,
-        rank=rank_int,
-        max_power_ori=max_power_ori,
-        inversion=inversion,
-    )
-
-    return filters
+    return out_nii_fname
 
 
-def _compute_beamformer(
-    G,
-    Cm,
-    reg,
-    n_orient,
-    weight_norm,
-    pick_ori,
-    reduce_rank,
-    rank,
-    inversion,
-    nn,
-    orient_std,
-    whitener,
+def recon_timeseries2niftii(
+    subjects_dir,
+    subject,
+    recon_timeseries,
+    out_nii_fname,
+    spatial_resolution=None,
+    reference_brain="mni",
+    times=None,
 ):
-    """Compute a spatial beamformer filter (LCMV or DICS).
-
-    For more detailed information on the parameters, see the docstrings of
-    `make_lcmv` and `make_dics`.
-
-    RHINO version of mne.beamformer._compute_beamformer
-    See lines marked MWW for where code has been changed
+    """Converts a (ndipoles,tpts) array of reconstructed timeseries (in
+    head/polhemus space) to the corresponding dipoles in a standard brain
+    grid in MNI space and outputs them as a niftii file.
 
     Parameters
     ----------
-    G : ndarray, shape (n_dipoles, n_channels)
-        The leadfield.
-    Cm : ndarray, shape (n_channels, n_channels)
-        The data covariance matrix.
-    reg : float
-        Regularization parameter.
-    n_orient : int
-        Number of dipole orientations defined at each source point
-    weight_norm : None | 'unit-noise-gain' | 'nai'
-        The weight normalization scheme to use.
-    pick_ori : None | 'normal' | 'max-power' | max-power-pre-weight-norm
-        The source orientation to compute the beamformer in.
-    reduce_rank : bool
-        Whether to reduce the rank by one during computation of the filter.
-    rank : dict | None | 'full' | 'info'
-        See compute_rank.
-    inversion : 'matrix' | 'single'
-        The inversion scheme to compute the weights.
-    nn : ndarray, shape (n_dipoles, 3)
-        The source normals.
-    orient_std : ndarray, shape (n_dipoles,)
-        The std of the orientation prior used in weighting the lead fields.
-    whitener : ndarray, shape (n_channels, n_channels)
-        The whitener.
+    subjects_dir : string
+        Directory to find RHINO subject dirs in.
+    subject : string
+        Subject name dir to find RHINO files in.
+    recon_timeseries : (ndipoles, ntpts) np.array
+        Reconstructed time courses (in head (polhemus) space).
+        Assumes that the dipoles are the same (and in the same order)
+        as those in the forward model,
+        coreg_filenames['forward_model_file'].
+        Typically derive from the VolSourceEstimate's output by
+        MNE source recon methods, e.g. mne.beamformer.apply_lcmv, obtained
+        using a forward model generated by RHINO.
+    spatial_resolution : int
+        Resolution to use for the reference brain in mm
+        (must be an integer, or will be cast to nearest int)
+        If None, then the gridstep used in coreg_filenames['forward_model_file']
+        is used.
+    reference_brain : string, 'mni' or 'mri'
+        'mni' indicates that the reference_brain is the stdbrain in MNI space
+        'mri' indicates that the reference_brain is the sMRI in native/mri space
+    times : (ntpts, ) np.array
+        Times points in seconds.
+        Will assume that these are regularly spaced
 
     Returns
     -------
-    W : ndarray, shape (n_dipoles, n_channels)
-        The beamformer filter weights.
-    """
-    _check_option(
-        "weight_norm",
-        weight_norm,
-        ["unit-noise-gain-invariant", "unit-noise-gain", "nai", None],
-    )
-
-    # Whiten the data covariance
-    Cm = whitener @ Cm @ whitener.T.conj()
-    # Restore to properly Hermitian as large whitening coefs can have bad
-    # rounding error
-
-    Cm[:] = (Cm + Cm.T.conj()) / 2.0
-
-    assert Cm.shape == (G.shape[0],) * 2
-    s, _ = np.linalg.eigh(Cm)
-    if not (s >= -s.max() * 1e-7).all():
-        # This shouldn't ever happen, but just in case
-        warn(
-            "data covariance does not appear to be positive semidefinite, "
-            "results will likely be incorrect"
-        )
-    # Tikhonov regularization using reg parameter to control for
-    # trade-off between spatial resolution and noise sensitivity
-    # eq. 25 in Gross and Ioannides, 1999 Phys. Med. Biol. 44 2081
-    Cm_inv, loading_factor, rank = _reg_pinv(Cm, reg, rank)
-
-    assert orient_std.shape == (G.shape[1],)
-    n_sources = G.shape[1] // n_orient
-    assert nn.shape == (n_sources, 3)
-
-    logger.info(
-        "Computing beamformer filters for %d source%s" % (n_sources, _pl(n_sources))
-    )
-    n_channels = G.shape[0]
-    assert n_orient in (3, 1)
-    Gk = np.reshape(G.T, (n_sources, n_orient, n_channels)).transpose(0, 2, 1)
-    assert Gk.shape == (n_sources, n_channels, n_orient)
-    sk = np.reshape(orient_std, (n_sources, n_orient))
-    del G, orient_std
-    pinv_kwargs = dict()
-    if check_version("numpy", "1.17"):
-        pinv_kwargs["hermitian"] = True
-
-    _check_option("reduce_rank", reduce_rank, (True, False))
-
-    # inversion of the denominator
-    _check_option("inversion", inversion, ("matrix", "single"))
-    if (
-        inversion == "single"
-        and n_orient > 1
-        and pick_ori == "vector"
-        and weight_norm == "unit-noise-gain-invariant"
-    ):
-        raise ValueError(
-            'Cannot use pick_ori="vector" with inversion="single" and '
-            'weight_norm="unit-noise-gain-invariant"'
-        )
-    if reduce_rank and inversion == "single":
-        raise ValueError(
-            'reduce_rank cannot be used with inversion="single"; '
-            'consider using inversion="matrix" if you have a '
-            "rank-deficient forward model (i.e., from a sphere "
-            "model with MEG channels), otherwise consider using "
-            "reduce_rank=False"
-        )
-    if n_orient > 1:
-        _, Gk_s, _ = np.linalg.svd(Gk, full_matrices=False)
-        assert Gk_s.shape == (n_sources, n_orient)
-        if not reduce_rank and (Gk_s[:, 0] > 1e6 * Gk_s[:, 2]).any():
-            raise ValueError(
-                "Singular matrix detected when estimating spatial filters. "
-                "Consider reducing the rank of the forward operator by using "
-                "reduce_rank=True."
-            )
-        del Gk_s
-
-    #
-    # 1. Reduce rank of the lead field
-    #
-    if reduce_rank:
-        Gk = _reduce_leadfield_rank(Gk)
-
-    def _compute_bf_terms(Gk, Cm_inv):
-        bf_numer = np.matmul(Gk.swapaxes(-2, -1).conj(), Cm_inv)
-        bf_denom = np.matmul(bf_numer, Gk)
-        return bf_numer, bf_denom
-
-    #
-    # 2. Reorient lead field in direction of max power or normal
-    #
-    if pick_ori == "max-power" or pick_ori == "max-power-pre-weight-norm":
-        assert n_orient == 3
-        _, bf_denom = _compute_bf_terms(Gk, Cm_inv)
-
-        if pick_ori == "max-power":
-            if weight_norm is None:
-                ori_numer = np.eye(n_orient)[np.newaxis]
-                ori_denom = bf_denom
-            else:
-                # compute power, cf Sekihara & Nagarajan 2008, eq. 4.47
-                ori_numer = bf_denom
-                # Cm_inv should be Hermitian so no need for .T.conj()
-                ori_denom = np.matmul(
-                    np.matmul(Gk.swapaxes(-2, -1).conj(), Cm_inv @ Cm_inv), Gk
-                )
-
-            ori_denom_inv = _sym_inv_sm(ori_denom, reduce_rank, inversion, sk)
-            ori_pick = np.matmul(ori_denom_inv, ori_numer)
-
-        # MWW
-        else:  # pick_ori == 'max-power-pre-weight-norm':
-
-            # Compute power, see eq 5 in Brookes et al, Optimising experimental design for MEG beamformer imaging, Neuroimage 2008
-            # This optimises the orientation by maximising the power
-            # BEFORE any weight normalisation is performed
-            ori_pick = _sym_inv_sm(bf_denom, reduce_rank, inversion, sk)
-
-        assert ori_pick.shape == (n_sources, n_orient, n_orient)
-
-        # pick eigenvector that corresponds to maximum eigenvalue:
-        eig_vals, eig_vecs = np.linalg.eig(ori_pick.real)  # not Hermitian!
-        # sort eigenvectors by eigenvalues for picking:
-        order = np.argsort(np.abs(eig_vals), axis=-1)
-        # eig_vals = np.take_along_axis(eig_vals, order, axis=-1)
-        max_power_ori = eig_vecs[np.arange(len(eig_vecs)), :, order[:, -1]]
-        assert max_power_ori.shape == (n_sources, n_orient)
-
-        # set the (otherwise arbitrary) sign to match the normal
-        signs = np.sign(np.sum(max_power_ori * nn, axis=1, keepdims=True))
-        signs[signs == 0] = 1.0
-        max_power_ori *= signs
-
-        # Compute the lead field for the optimal orientation,
-        # and adjust numer/denom
-
-        Gk = np.matmul(Gk, max_power_ori[..., np.newaxis])
-
-        n_orient = 1
-    else:
-        max_power_ori = None
-        if pick_ori == "normal":
-            Gk = Gk[..., 2:3]
-            n_orient = 1
-
-    #
-    # 3. Compute numerator and denominator of beamformer formula (unit-gain)
-    #
-
-    bf_numer, bf_denom = _compute_bf_terms(Gk, Cm_inv)
-    assert bf_denom.shape == (n_sources,) + (n_orient,) * 2
-    assert bf_numer.shape == (n_sources, n_orient, n_channels)
-    del Gk  # lead field has been adjusted and should not be used anymore
-
-    #
-    # 4. Invert the denominator
-    #
-
-    # Here W is W_ug, i.e.:
-    # G.T @ Cm_inv / (G.T @ Cm_inv @ G)
-    bf_denom_inv = _sym_inv_sm(bf_denom, reduce_rank, inversion, sk)
-    assert bf_denom_inv.shape == (n_sources, n_orient, n_orient)
-    W = np.matmul(bf_denom_inv, bf_numer)
-    assert W.shape == (n_sources, n_orient, n_channels)
-    del bf_denom_inv, sk
-
-    #
-    # 5. Re-scale filter weights according to the selected weight_norm
-    #
-
-    # Weight normalization is done by computing, for each source::
-    #
-    #     W_ung = W_ug / sqrt(W_ug @ W_ug.T)
-    #
-    # with W_ung referring to the unit-noise-gain (weight normalized) filter
-    # and W_ug referring to the above-calculated unit-gain filter stored in W.
-
-    if weight_norm is not None:
-        # Three different ways to calculate the normalization factors here.
-        # Only matters when in vector mode, as otherwise n_orient == 1 and
-        # they are all equivalent. Sekihara 2008 says to use
-        #
-        # In MNE < 0.21, we just used the Frobenius matrix norm:
-        #
-        #    noise_norm = np.linalg.norm(W, axis=(1, 2), keepdims=True)
-        #    assert noise_norm.shape == (n_sources, 1, 1)
-        #    W /= noise_norm
-        #
-        # Sekihara 2008 says to use sqrt(diag(W_ug @ W_ug.T)), which is not
-        # rotation invariant:
-        if weight_norm in ("unit-noise-gain", "nai"):
-            noise_norm = np.matmul(W, W.swapaxes(-2, -1).conj()).real
-            noise_norm = np.reshape(  # np.diag operation over last two axes
-                noise_norm, (n_sources, -1, 1)
-            )[:, :: n_orient + 1]
-            np.sqrt(noise_norm, out=noise_norm)
-            noise_norm[noise_norm == 0] = np.inf
-            assert noise_norm.shape == (n_sources, n_orient, 1)
-            W /= noise_norm
-        else:
-            assert weight_norm == "unit-noise-gain-invariant"
-            # Here we use sqrtm. The shortcut:
-            #
-            #    use = W
-            #
-            # ... does not match the direct route (it is rotated!), so we'll
-            # use the direct one to match FieldTrip:
-            use = bf_numer
-            inner = np.matmul(use, use.swapaxes(-2, -1).conj())
-            W = np.matmul(_sym_mat_pow(inner, -0.5), use)
-            noise_norm = 1.0
-
-        if weight_norm == "nai":
-            # Estimate noise level based on covariance matrix, taking the
-            # first eigenvalue that falls outside the signal subspace or the
-            # loading factor used during regularization, whichever is largest.
-            if rank > len(Cm):
-                # Covariance matrix is full rank, no noise subspace!
-                # Use the loading factor as noise ceiling.
-                if loading_factor == 0:
-                    raise RuntimeError(
-                        "Cannot compute noise subspace with a full-rank "
-                        "covariance matrix and no regularization. Try "
-                        "manually specifying the rank of the covariance "
-                        "matrix or using regularization."
-                    )
-                noise = loading_factor
-            else:
-                noise, _ = np.linalg.eigh(Cm)
-                noise = noise[-rank]
-                noise = max(noise, loading_factor)
-            W /= np.sqrt(noise)
-
-    W = W.reshape(n_sources * n_orient, n_channels)
-    logger.info("Filter computation complete")
-    return W, max_power_ori
-
-
-def _prepare_beamformer_input(
-    info,
-    forward,
-    label=None,
-    pick_ori=None,
-    noise_cov=None,
-    rank=None,
-    pca=False,
-    loose=None,
-    combine_xyz="fro",
-    exp=None,
-    limit=None,
-    allow_fixed_depth=True,
-    limit_depth_chs=False,
-):
-    """Input preparation common for LCMV, DICS, and RAP-MUSIC.
-    RHINO version of mne.beamformer._prepare_beamformer_input
-    See lines marked MWW for where code has been changed
+    out_nii_fname : string
+        Name of output niftii file
+    reference_brain_fname : string
+        Niftii file name of standard brain mask in MNI space at requested resolution,
+        int(stdbrain_resolution)
+        (with zero for background, and !=0 for the mask)
     """
 
-    # MWW
-    # _check_option('pick_ori', pick_ori,
-    #              ('normal', 'max-power', 'vector', None))
-    _check_option(
-        "pick_ori",
-        pick_ori,
-        ("normal", "max-power", "vector", "max-power-pre-weight-norm", None),
+    if len(recon_timeseries.shape) == 1:
+        recon_timeseries = np.reshape(recon_timeseries, [recon_timeseries.shape[0], 1])
+
+    # ---------------------------------------------------
+    # convert the recon_timeseries to the standard
+    # space brain dipole grid at the specified resolution
+    (
+        recon_ts_out,
+        reference_brain_fname,
+        recon_coords_out,
+        recon_indices,
+    ) = transform_recon_timeseries(
+        subjects_dir,
+        subject,
+        recon_timeseries=recon_timeseries,
+        spatial_resolution=spatial_resolution,
+        reference_brain=reference_brain,
     )
 
-    # Restrict forward solution to selected vertices
-    if label is not None:
-        _, src_sel = label_src_vertno_sel(label, forward["src"])
-        forward = _restrict_forward_to_src_sel(forward, src_sel)
+    # ----------------------------------
+    # output recon_ts_out as niftii file
+    out_nii_fname = _timeseries2nii(
+        recon_ts_out,
+        recon_coords_out,
+        reference_brain_fname,
+        out_nii_fname,
+        times=times,
+    )
 
-    if loose is None:
-        loose = 0.0 if is_fixed_orient(forward) else 1.0
-    if noise_cov is None:
-        noise_cov = make_ad_hoc_cov(info, std=1.0)
-    forward, info_picked, gain, _, orient_prior, _, trace_GRGT, noise_cov, whitener = \
-        _prepare_forward(
-            forward, info, noise_cov, "auto", loose, rank=rank, pca=pca, use_cps=True,
-            exp=exp, limit_depth_chs=limit_depth_chs, combine_xyz=combine_xyz,
-            limit=limit, allow_fixed_depth=allow_fixed_depth,
-        )
-    is_free_ori = not is_fixed_orient(forward)  # could have been changed
-    nn = forward["source_nn"]
-    if is_free_ori:  # take Z coordinate
-        nn = nn[2::3]
-    nn = nn.copy()
-    vertno = _get_vertno(forward["src"])
-    if forward["surf_ori"]:
-        nn[...] = [0, 0, 1]  # align to local +Z coordinate
-    if pick_ori is not None and not is_free_ori:
-        raise ValueError(
-            "Normal or max-power orientation (got %r) can only be picked when "
-            "a forward operator with free orientation is used." % (pick_ori,)
-        )
-    if pick_ori == "normal" and not forward["surf_ori"]:
-        raise ValueError(
-            "Normal orientation can only be picked when a "
-            "forward operator oriented in surface coordinates is "
-            "used."
-        )
-    _check_src_normal(pick_ori, forward["src"])
-    del forward, info
-
-    # Undo the scaling that MNE prefers
-    scale = np.sqrt((noise_cov["eig"] > 0).sum() / trace_GRGT)
-    gain /= scale
-    if orient_prior is not None:
-        orient_std = np.sqrt(orient_prior)
-    else:
-        orient_std = np.ones(gain.shape[1])
-
-    # Get the projector
-    proj, _, _ = make_projector(info_picked["projs"], info_picked["ch_names"])
-    return (is_free_ori, info_picked, proj, vertno, gain, whitener, nn, orient_std)
+    return out_nii_fname, reference_brain_fname
 
 
 def save_or_show_renderer(renderer, filename):
